touch: gt9xx: using delayed workqueu instead of hrtimer
The current code of polling mode does not work as the code accessing I2C
does not work in interrupt mode. The code work only in non-interrupt
context. So if we need to use the timer, we have to schedule a
workqueue. In this case, we will used delayed workqueue directly instead
of using hrtimer plus workqueue.

Further changes are needed for interrupt mode to make it work.